% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesearch_helpers.R
\name{biter}
\alias{biter}
\title{fits pls models at each \emph{gesearch} iteration}
\usage{
biter(
  itersubs,
  Xu,
  Yu,
  y_lim_left,
  y_lim_right,
  iter_sequence,
  n,
  optimization,
  ncomp,
  tune,
  p,
  number,
  scale,
  max_iter,
  tol,
  seed = NULL,
  algorithm,
  allow_parallel,
  pchunks = 1
)
}
\description{
internal
}
\keyword{internal}
