% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_projection_helpers.R
\name{check_pc_arguments}
\alias{check_pc_arguments}
\title{checks the pc_selection argument}
\usage{
check_pc_arguments(
  n_rows_x,
  n_cols_x,
  pc_selection,
  default_max_comp = 40,
  default_max_cumvar = 0.99,
  default_max_var = 0.01
)
}
\description{
internal
}
\keyword{internal}
