% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cms}
\alias{cms}
\title{Function for computing the mean of each column in a \code{matrix}}
\usage{
colSds(X)
}
\arguments{
\item{X}{a a \code{matrix}.}
}
\value{
a vector of mean values.
}
\description{
Computes the mean of each column in a \code{matrix}. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

