% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colSds}
\alias{colSds}
\title{Function for computing the standard deviation of each column in a \code{matrix}}
\usage{
colSds(X)
}
\arguments{
\item{X}{a a \code{matrix}.}
}
\value{
a vector of standard deviation values.
}
\description{
Computes the standard deviation of each column in a \code{matrix}. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

