% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cSums}
\alias{cSums}
\title{Function for computing sum of each column in a \code{matrix}}
\usage{
cSums(X)
}
\arguments{
\item{X}{a \code{matrix}.}
}
\value{
a vector of standard deviation values.
}
\description{
Computes the sum of each column in a \code{matrix}. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
