% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dc_svd}
\alias{dc_svd}
\title{Singular value decomposition (SVD)}
\usage{
dc_svd(X)
}
\arguments{
\item{X}{a \code{matrix}.}
}
\description{
Computes the SVD based on the divide and conquer algorithm
}
\author{
Leonardo Ramirez-Lopez based on code from Dirk Eddelbuettel
}
\keyword{internal}
