% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diss_methods.R
\name{diss_euclidean}
\alias{diss_euclidean}
\title{Euclidean dissimilarity method constructor}
\usage{
diss_euclidean(center = TRUE, scale = FALSE)
}
\arguments{
\item{center}{Logical. Center the data before computing distances?
Applied jointly to \code{Xr} and \code{Xu} if both are provided.
Default \code{TRUE}.}

\item{scale}{Logical. Scale the data before computing distances?
Applied jointly to \code{Xr} and \code{Xu} if both are provided.
Default \code{FALSE}.}
}
\value{
An object of class \code{c("diss_euclidean", "diss_method")}.
}
\description{
\loadmathjax
Creates a configuration object for computing Euclidean dissimilarity.
Pass the result to \code{dissimilarity()} to compute the dissimilarity
matrix.

The scaled Euclidean dissimilarity between two observations \mjeqn{x_i}{x_i} and
\mjeqn{x_j}{x_j} is:

\mjdeqn{d(x_i, x_j) = \sqrt{\frac{1}{p} \sum_{k=1}^{p}(x_{i,k} - x_{j,k})^2}}{d(x_i, x_j) = sqrt(1/p * sum_k (x_i,k - x_j,k)^2)}

where \mjeqn{p}{p} is the number of variables. Results are equivalent to
\code{stats::dist()} but scaled by \mjeqn{1/p}{1/p}.
}
\examples{
m <- diss_euclidean()
m <- diss_euclidean(center = FALSE, scale = TRUE)
}
\seealso{
\code{\link{dissimilarity}}, \code{\link{diss_mahalanobis}},
\code{\link{diss_cosine}}
}
\author{
\href{https://orcid.org/0000-0002-5369-5120}{Leonardo Ramirez-Lopez}
}
