% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesearch_helpers.R
\name{drop_indices}
\alias{drop_indices}
\title{drop the samples that consistently appear in high error models}
\usage{
drop_indices(
  uu,
  vv,
  r,
  obs_indices,
  resampling_indices,
  weakness_subset,
  retain_by,
  cull_quantity,
  percentile_type,
  max_pls_c
)
}
\description{
internal
}
\keyword{internal}
