% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fDiss.R
\name{e2m}
\alias{e2m}
\title{A function for transforming a matrix from its Euclidean spcae to its Mahalanobis space}
\usage{
e2m(X, sm.method = c("svd", "eigen"))
}
\description{
For internal use only
}
\keyword{internal}

