% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_diss.R
\name{euclid_to_mahal}
\alias{euclid_to_mahal}
\title{A function for transforming a matrix from its Euclidean space to 
its Mahalanobis space}
\usage{
euclid_to_mahal(X, sm_method = c("svd", "eigen"))
}
\description{
For internal use only
}
\keyword{internal}
