% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_projection_helpers.R
\name{eval_multi_pc_diss}
\alias{eval_multi_pc_diss}
\title{evaluation of multiple distances obtained with multiple PCs}
\usage{
eval_multi_pc_diss(
  scores,
  side_info,
  from = 1,
  to = ncol(scores),
  steps = 1,
  method = c("pc", "pls"),
  check_dims = TRUE
)
}
\description{
internal
}
\keyword{internal}
