% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_diss}
\alias{fast_diss}
\title{A fast distance algorithm for two matrices written in C++}
\usage{
fast_diss(X, Y, method)
}
\arguments{
\item{X}{a matrix}

\item{Y}{a matrix}

\item{method}{a \code{string} with possible values "euclid", "cor", "cosine"}
}
\value{
a distance matrix
}
\description{
Computes distances between two data matrices using
"euclid", "cor", "cosine"
}
\author{
Antoine Stevens and Leonardo Ramirez-Lopez
}
\keyword{internal}
