% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_diss_vector}
\alias{fast_diss_vector}
\title{A fast algorithm of (squared) Euclidean cross-distance for vectors written in C++}
\usage{
fast_diss_vector(X)
}
\arguments{
\item{X}{a vector.}
}
\value{
a vector of distance (lower triangle of the distance matrix, stored by column)
}
\description{
A fast (parallel for linux) algorithm of (squared) Euclidean cross-distance for vectors written in C++
}
\details{
used internally in ortho_projection
}
\author{
Antoine Stevens
}
\keyword{internal}
