% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastDistVVL}
\alias{fastDistVVL}
\title{A fast (serial) algorithm of Euclidean (non-squared) cross-distance for vectors written in C++}
\usage{
fastDistVVL(X)
}
\arguments{
\item{X}{a vector}
}
\value{
a vector of distance (lower triangle of the distance matrix, stored by column)
}
\description{
A fast algorithm of (squared) Euclidean cross-distance for vectors written in C++
}
\details{
used internally in ortho_projection
}
\author{
Leo Ramirez-Lopez
}
\keyword{internal}
