% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_helpers.R
\name{fit_and_predict}
\alias{fit_and_predict}
\title{Local multivariate regression}
\usage{
fit_and_predict(
  x,
  y,
  pred_method,
  scale = FALSE,
  weights = NULL,
  newdata,
  pls_c = NULL,
  CV = FALSE,
  tune = FALSE,
  number = 10,
  p = 0.75,
  group = NULL,
  noise_variance = 0.001,
  range_prediction_limits = TRUE,
  pls_max_iter = 1,
  pls_tol = 1e-06,
  modified = FALSE
)
}
\description{
internal
}
\keyword{internal}
