% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_library.R
\name{fit_library}
\alias{fit_library}
\title{Build predictive/fucntional libraries}
\usage{
fit_library(Xr, Yr, sm, k, k_diss, k_range, ws = NULL, pls_c, 
            pc_selection = list("opc", min(dim(Xr), 40)),
            group, build_around, gh = TRUE, center = TRUE,
            scale = FALSE, diss_predictors = FALSE,
            metric = "rmse", return_best = TRUE,
            pls_max_iter = 1, pls_tol = 1e-6,
            documentation = character(), ...)
}
\arguments{
\item{Xr}{a numeric matrix (or data.frame) of predictor variables of dimentions \code{n * p} corresponding to the reference data (observations in rows and variables in columns).}

\item{Yr}{a numeric vector of length \code{n} containing the values of the response variable corresponding to the reference data. Missing values might be allowed (see details).}

\item{diss_method}{a character string indicating the spectral dissimilarity metric to be used in the selection of the nearest neighbours of each observation.
Options are:
\itemize{
\item{\code{"pc"}: Principal components dissimilarity: Mahalanobis dissimilarity computed on the principal components space.}
\item{\code{"cor"}: Correlation dissimilarity.}
\item{\code{"pls"}: Partial least squares dissimilarity: Mahalanobis dissimilarity computed on the partial least squares space.}
\item{\code{"euclid"}: Euclidean dissimilarity.}
\item{\code{"cosine"}: Cosine dissimilarity.}
\item{\code{"sidF"}: Spectral information divergence computed on the spectral variables.}
\item{\code{"sidD"}: Spectral information divergence computed on the density distributions of the spectra.}
\item{\code{"loc.pc"}: Dissimilarity estimation based on local principal components.}
\item{\code{"loc.pls"} Dissimilarity estimation based on local partial least squares.}
}
The \code{"pc"} spectral dissimilarity metric is the default. If the \code{"sidD"} is chosen, the default parameters of the \code{sid} function are used however they can be specified by passing them them as additional arguments \code{...}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{   ======= NOT YET IMPLEMENTED!!! ======= --> --> --> Alternatively a squared dissimilarity matrix of size `n * n` can also be passed. In this case, each row of the matrix must correspond to the observations. The nearest neighbors search is conducted column-wise.
}\if{html}{\out{</div>}}}

\item{k}{a vector of integers specifying the sequence of k nearest neighbours to be tested. Either \code{k} or \code{k_diss} must be specified. This \code{vector} will be automatically sorted into ascending order. The values of vectors of class numeric, will be coerced to the next upper inetegers.}

\item{k_diss}{NOT YET IMPLEMENTED a numeric \code{vector} specifying the sequence of dissimilarity thresholds to be tested for the selection of the nearest neighbors around each observation. For a given observation, its neighbors are those that exhibit a dissimilarity value equal to or below a given threshold. These thresholds depend on the corresponding dissimilarity measure specified in \code{sm}. Either \code{k} or \code{k_diss} must be specified.}

\item{k_range}{NOT YET IMPLEMENTED a vector of two integers specifying the minimum (first value) and the maximum (second value) number of neighbours allowed when the \code{k_diss} argument is used.}

\item{ws}{an optional odd integer value which specifies a window size when the correlation dissimilarity is used (i.e \code{sm = "cor"}). If not specified, the standard correlation dissimilarity is computed (i.e no moving window). Default is \code{NULL}.}

\item{pls_c}{a vector of two integers indicating the minimum (first value) and maximum (second value) number of pls components to be used in the weighted average pls (wapls) regressions.}

\item{pc_selection}{a list specifying the details of the method to be used for identifying the number of components to be retained for computing the dissimilarities between samples when the method passed to \code{sm} is one of the following options: \code{"pc"}, \code{"loc.pc"}, \code{ "pls"} or \code{ "loc.pls"}. This list must contain two elements in the following order: \itemize{
\item{\code{method}:}{the method for selecting the number of components. Possible options are:  \code{"opc"} (optimized pc selection based on Ramirez-Lopez et al. (2013a, 2013b). See the \code{\link{orthoProjection}} function for more details;  \code{"cumvar"} (for selecting the number of principal components based on a given cumulative amount of explained variance); \code{"var"} (for selecting the number of principal components based on a given amount of explained variance); and  \code{"manual"} (for specifying manually the desired number of principal components)}
\item{\code{value}:}{a numerical value that complements the selected method. If \code{"opc"} is chosen, it must be a value indicating the maximal number of principal components to be tested (see Ramirez-Lopez et al., 2013a, 2013b). If \code{"cumvar"} is chosen, it must be a value (larger than 0 and below 1) indicating the maximum amount of cumulative variance that the retained components should explain. If \code{"var"} is chosen, it must be a value (larger than 0 and below 1) indicating that components that explain (individually) a variance lower than this threshold must be excluded. If \code{"manual"} is chosen, it must be a value specifying the desired number of principal components to retain.
}}
The default method for the \code{pc_selection} argument is \code{"opc"} and the maximum number of principal components to be tested is set to \code{min(dim(Xr), 40)}, where code{Xr} is the matrix of reference predictors in the \code{\link{mbl}} function.
Optionally, the \code{pc_selection} argument admits \code{"opc"} or \code{"cumvar"} or \code{"var"} or \code{"manual"} as a single character string. In such a case the default for \code{"value"} when either \code{"opc"} or \code{"manual"} are used is \code{min(dim(Xr), 40)}. When \code{"cumvar"} is used the default \code{"value"} is set to 0.99 and when \code{"var"} is used the default \code{"value"} is set to 0.01.}

\item{group}{(PUT THE DETAILS OF THE 1_NN VALIDATION IN DETAILS) an optional \code{factor} (or \code{vector} that can be coerced to a \code{factor} by \code{as.factor}) that assigns to each observation in \code{Xr} a group/class label (e.g. groups can be given by spectra collected from the same batch of measurements, from the same sample, from samples with very similar origin, etc). This is taken into account for internal validation of pls models (and factor optimization) to avoid pseudo-replication. For validation, the model of a given target observation is first fitted based on its neighbors and excluding that observation (and all the samples belonging to its group), then the model is used to predict the response variable of the target observation. See details.}

\item{build_around}{NOT YET IMPLEMENTED! a vector of integers of length \code{m} (\code{m < n}) indicating the indices of the samples in the reference set for which the local models must be fitted. Default is \code{NULL} which means that local models are fitted for all the \code{n} observations in the reference set. See details.}

\item{gh}{a logical indicating whether or not to compute and return the Mahalanobis distance (in the pls space) between each element in \code{Xr} and the center of \code{Xr}.}

\item{center}{a logical indicating whether or not the predictor variables must be centered at each local segment (before regression).}

\item{scale}{a logical indicating whether or not the predictor variables must be scaled to unit variance at each local segment (before regression).}

\item{diss_predictors}{a logical indicating if the local (square symmetric) dissimilarity matrix corresponding the selected neighbourhood shall be used as source of additional predictors (i.e the columns of this local matrix are treated as predictor variables). In some cases this may result in an improvement  of the prediction performance (Ramirez-Lopez et al., 2013a).}

\item{metric}{a character value indicating what model perfoamance statistics shall be used to select the best set of parameters (i.e. number of neighbors or threshold distances and pls factors)  to fit the final model. Options are \code{"rmse"} (default) or \code{"r2"} (coefficient of determination).}

\item{return_best}{a logical indicating if the final library of functions using the optimal parameters found shall be returned.}

\item{pls_max_iter}{(BETTER DESCRIPTION REQUIRED) maximum number of iterations for the partial least squares methods.}

\item{pls_tol}{(BETTER DESCRIPTION REQUIRED) for convergence in the partial orthogonal scores partial least squares regressions using the nipals algorithm. Default is 1e-6}

\item{documentation}{(BETTER DESCRIPTION REQUIRED) an optional character string for documentating the call to this function.}

\item{...}{arguments passed to the \code{\link{dissimilarity}} function.}
}
\value{
DESCRIOTION REQUIRED
}
\description{
This function builds a library of predictive functions based on memory-based learning
}
\details{
By default, this function fits a local model for each of the \code{n} observations in the reference set, i.e. a library of \code{n} fucntions is built. Each local model is fitted with the query observation and its nearest neighbors.
Alternatively, the user may specificy the samples in the reference set around which the library of functions must be built. This is done by indicating in the \code{build_around} argument the indices of the samples in the reference set. In tis case, each function is fitted by using the nearest neighbors found in the entire set of \code{n} reference observations. The \code{build_around} argument may be useful in cases where the number of observations in the reference set is extremely large.
Missing values in \code{Yr} are allowed. If they exceed 25\% of the total number of observations a warning message will be printed. The local model of an observation with missing \code{Yr} value is fited only with its nearest neighbors (i.e without including this observation).
}
\examples{
\dontrun{
#' ## GOOD EXAMPLES ARE REQUIRED
}
}
\references{
Rajalahti, T., Arneberg, R., Berven, F. S., Myhr, K. M., Ulvik, R. J., & Kvalheim, O. M. (2009). Biomarker discovery in mass spectral profiles by means of selectivity ratio plot. Chemometrics and Intelligent Laboratory Systems, 95(1), 35-48.
}
\author{
Leonardo Ramirez-Lopez
}
