% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_diss_helpers.R
\name{format_xr_xu_indices}
\alias{format_xr_xu_indices}
\title{format internal messages}
\usage{
format_xr_xu_indices(xr_xu_names)
}
\arguments{
\item{xr_xu_names}{the names of Xr and Xu}
}
\description{
internal
}
\keyword{internal}
