% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_helpers.R
\name{gaussian_pr_cv}
\alias{gaussian_pr_cv}
\title{Cross validation for Gaussian process regression}
\usage{
gaussian_pr_cv(
  x,
  y,
  scale,
  weights = NULL,
  p = 0.75,
  number = 10,
  group = NULL,
  noise_variance = 0.001,
  retrieve = c("final_model", "none"),
  seed = NULL
)
}
\description{
internal
}
\keyword{internal}
