% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_col_largest_sd}
\alias{get_col_largest_sd}
\title{Function for identifiying the column in a matrix with the largest standard deviation}
\usage{
get_col_largest_sd(X)
}
\arguments{
\item{X}{a matrix.}
}
\value{
a value indicating the index of the column with the largest standard deviation.
}
\description{
Identifies the column with the largest standard deviation. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
