% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_column_maxs}
\alias{get_column_maxs}
\title{Function for computing the maximum of each column}
\usage{
get_column_maxs(X)
}
\arguments{
\item{X}{a a matrix.}
}
\value{
a vector of mean values.
}
\description{
Computes the max of each column in a matrix. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
