% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_column_means}
\alias{get_column_means}
\title{Function for computing the mean of each column in a matrix}
\usage{
get_column_means(X)
}
\arguments{
\item{X}{a a matrix.}
}
\value{
a vector of mean values.
}
\description{
Computes the mean of each column in a matrix. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
