% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_column_sds}
\alias{get_column_sds}
\title{Function for computing the standard deviation of each column in a matrix}
\usage{
get_column_sds(X)
}
\arguments{
\item{X}{a a matrix.}
}
\value{
a vector of standard deviation values.
}
\description{
Computes the standard deviation of each column in a matrix. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
