% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_column_sums}
\alias{get_column_sums}
\title{Function for computing sum of each column in a matrix}
\usage{
get_column_sums(X)
}
\arguments{
\item{X}{a matrix.}
}
\value{
a vector of standard deviation values.
}
\description{
Computes the sum of each column in a matrix. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
