% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_local_pls_weights}
\alias{get_local_pls_weights}
\title{Internal Cpp function for computing the weights of the PLS components
necessary for weighted average PLS}
\usage{
get_local_pls_weights(projection_mat, 
          xloadings, 
          coefficients, 
          new_x, 
          min_component, 
          max_component, 
          scale, 
          Xcenter, 
          Xscale)
}
\arguments{
\item{projection_mat}{the projection matrix generated either by the \code{opls} function.}

\item{xloadings}{.}

\item{coefficients}{the matrix of regression coefficients.}

\item{new_x}{a matrix of one new spectra to be predicted.}

\item{min_component}{an integer indicating the minimum number of pls components.}

\item{max_component}{an integer indicating the maximum number of pls components.}

\item{scale}{a logical indicating whether the matrix of predictors used to create the regression model was scaled.}

\item{Xcenter}{a vector with the values that must be used for centering \code{newdata}.}

\item{Xscale}{if \code{scale = TRUE} a vector with the values that must be used for scaling \code{newdata}.}
}
\value{
a matrix of one row with the weights for each component between the max. and min. specified.
}
\description{
For internal use only!.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
