% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_helpers.R
\name{get_neighbor_info}
\alias{get_neighbor_info}
\title{A function to get the neighbor information}
\usage{
get_neighbor_info(
  Xr,
  Xu,
  diss_method,
  Yr = NULL,
  k = NULL,
  k_diss = NULL,
  k_range = NULL,
  spike = NULL,
  pc_selection,
  return_dissimilarity,
  center,
  scale,
  gh,
  diss_usage,
  allow_parallel = FALSE,
  ...
)
}
\description{
This fucntion gathers information of all neighborhoods of the 
\code{Xu} observations found in \code{Xr}. This information is equired during 
local regressions.
}
\details{
For local pca and pls distances, the local dissimilarity matrices are not 
computed as it is cheaer to compute them during the local regressions. 
Instead the global distances (required for later local dissimilarity matrix
computation are output)
}
\keyword{internal}
