% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predictions.R
\name{get_predictions}
\alias{get_predictions}
\title{Extract predictions from an object of class \code{mbl}}
\usage{
get_predictions(object)
}
\arguments{
\item{object}{an object of class \code{mbl} as returned by \code{mbl}}
}
\value{
a data.table of predicted values according to either \code{k} or \code{k_dist}
}
\description{
\lifecycle{stable}

Extract predictions from an object of class \code{mbl}
}
\seealso{
\code{\link{mbl}}
}
\author{
Leonardo Ramirez-Lopez and Antoine Stevens
}
