% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_str.R
\name{get_sample_strata}
\alias{get_sample_strata}
\title{A function to assign values to sample distribution strata}
\usage{
get_sample_strata(y, n)
}
\arguments{
\item{y}{a matrix of one column with the response variable.}

\item{n}{the number of strata.}
}
\value{
a data table with the input \code{y} and the corresponding strata to 
every value.
}
\description{
for internal use only! This function takes a continuous variable,
creates n strata based on its distribution and assigns the corresponding starta
to every value.
}
\keyword{internal}
