% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stratified.R
\name{get_samples_from_strata}
\alias{get_samples_from_strata}
\title{A function for stratified calibration/validation sampling}
\usage{
get_samples_from_strata(
  original_order,
  strata,
  samples_per_strata,
  sampling_for = c("calibration", "validation"),
  replacement = FALSE
)
}
\arguments{
\item{original_order}{a matrix of one column with the response variable.}

\item{sampling_for}{sampling to select the calibration samples ("calibration")
or sampling to select the validation samples ("validation").}

\item{replacement}{logical indicating if sampling with replacement must be
done.}

\item{starta}{the number of strata.}
}
\value{
a list with the indices of the calibration and validation samples.
}
\description{
for internal use only! This function selects samples
based on provided strata.
}
\keyword{internal}
