% Generated by roxygen2 (4.0.1): do not edit by hand
\name{gpr.dp}
\alias{gpr.dp}
\title{Gaussian process regression with dot product covariance}
\usage{
gpr.dp(Xn, Yn, noise.v = 0.001, scaled = TRUE, center = FALSE,
  x.add = NULL)
}
\description{
Gaussian process regression with dot product covariance
}
\keyword{internal}

