\name{gprCv}
\alias{gprCv}
\title{Cross validation for Gaussian process regression}
\usage{
  gprCv(x, y, scaled, weights = NULL, p = 0.75,
    resampling = 10, kpar, fit,
    retrieve = c("final.model", "none"))
}
\description{
  Cross validation for Gaussian process regression
}
\keyword{internal}

