% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_diss_helpers.R
\name{local_ortho_diss}
\alias{local_ortho_diss}
\title{local ortho dissimilarity matrices initialized by a global
dissimilarity matrix}
\usage{
local_ortho_diss(
  k_index_matrix,
  Xr,
  Yr,
  Xu,
  diss_method,
  pc_selection,
  center,
  scale,
  allow_parallel,
  ...
)
}
\arguments{
\item{k_index_matrix}{a matrix of nearest neighnbor indices}

\item{Xr}{argument passed to ortho_projection}

\item{Yr}{argument passed to ortho_projection}

\item{Xu}{argument passed to ortho_projection}

\item{diss_method}{argument passed to ortho_projection}

\item{pc_selection}{argument passed to ortho_projection}

\item{center}{argument passed to ortho_projection}

\item{scale}{argument passed to ortho_projection}
}
\description{
internal
}
\keyword{internal}
