% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_visNIR.R
\docType{data}
\name{local_visNIR}
\alias{local_visNIR}
\title{local_visNIR}
\format{
A `data.frame` of 30 observations and 3 variables (where the spectral data is embedded in one variable `local_visNIR$spc`).
}
\source{
Raphael Viscarra-Rossel
}
\usage{
data(local_visNIR)
}
\description{
Soil spectral library containing soil clay values in percent and the vis-NIR soil spectra.
The database contains absorbance spectra of dried and sieved soil samples measured between 350 nm
and 2500 nm at 2 nm interval. One soil property is associated with the spectral library: clay content in percent. Each spectrum has a unique id value.
}
\details{
No further details.
}
\references{
Rossel, R. V., & Behrens, T. (2010). Using data mining to model and interpret soil diffuse reflectance spectra. Geoderma, 158(1-2), 46-54.
}
\keyword{datasets}
