\name{locFit}
\alias{locFit}
\title{Local multivariate regression}
\usage{
  locFit(x, y, predMethod, scaled = TRUE, weights = NULL,
    pred.new = TRUE, newdata, plsC, CV = FALSE,
    resampling = 10, p = 0.75, w2_warning = TRUE)
}
\description{
  Local multivariate regression
}
\keyword{internal}

