% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl.R
\name{locFitnpred}
\alias{locFitnpred}
\title{Local multivariate regression}
\usage{
locFitnpred(
  x,
  y,
  predMethod,
  scaled = TRUE,
  weights = NULL,
  newdata,
  pls.c = NULL,
  CV = FALSE,
  .optimize = FALSE,
  number = 10,
  p = 0.75,
  group = NULL,
  noise.v = 0.001,
  range.pred.lim = TRUE,
  pls.max.iter = 1,
  pls.tol = 1e-06
)
}
\description{
internal
}
\keyword{internal}
