% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moving_cor_diss}
\alias{moving_cor_diss}
\title{Moving/rolling correlation distance of two matrices}
\usage{
moving_cor_diss(X,Y,w)
}
\arguments{
\item{X}{a matrix}

\item{Y}{a matrix}

\item{w}{window size (must be odd)}
}
\value{
a matrix of correlation distance
}
\description{
Computes a moving window correlation distance between two data matrices
}
\author{
Leonardo Ramirez-Lopez and Antoine Stevens
}
\keyword{internal}
