% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moving_cor_diss_self}
\alias{moving_cor_diss_self}
\title{Rolling correlation distance within X (templated, OpenMP; precision)}
\usage{
moving_cor_diss_self(X, w, block_rows = 1024L, precision = "double")
}
\arguments{
\item{X}{Numeric matrix (m×T)}

\item{w}{Window size (odd in \link{1..T} or exactly T)}

\item{block_rows}{Tile size (default 1024)}

\item{precision}{"double" (default) or "float32"}
}
\value{
m×m distance matrix (double for R)
}
\description{
Rolling correlation distance within X (templated, OpenMP; precision)
}
