% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opls_get_basics}
\alias{opls_get_basics}
\title{Fast orthogonal scores algorithm of partial least squares (PLS)}
\usage{
opls_get_basics(X, Y, ncomp, scale, 
                maxiter, tol, 
                algorithm = "pls", 
                xls_min_w = 3, 
                xls_max_w = 15)
}
\arguments{
\item{X}{a matrix of predictor variables.}

\item{Y}{a matrix of either a single or multiple response variables.}

\item{ncomp}{the number of PLS components.}

\item{scale}{logical indicating whether \code{X} must be scaled.}

\item{maxiter}{maximum number of iterations (only used for NIPALS-based
algorithms: \code{'pls'}, \code{'mpls'}, \code{'xls'}).}

\item{tol}{convergence tolerance for the NIPALS algorithm (only used for
NIPALS-based algorithms).}

\item{algorithm}{a character string indicating the PLS algorithm to use:
\itemize{
\item{\code{'pls'}: standard PLS using covariance between X and Y for
weight computation (NIPALS algorithm).}
\item{\code{'mpls'}: modified PLS using correlation between X and Y for
weight computation (NIPALS algorithm). See Shenk and Westerhaus (1991).}
\item{\code{'xls'}: extended PLS as implemented in BUCHI NIRWise PLUS
software (NIPALS algorithm).}
\item{\code{'simpls'}: SIMPLS algorithm (de Jong, 1993). Computationally
faster as it avoids iterative X deflation. Parameters \code{maxiter},
\code{tol}, \code{xls_min_w}, and \code{xls_max_w} are ignored.}
}}

\item{xls_min_w}{an integer indicating the minimum window size for the
\code{'xls'} method. Only used if \code{algorithm = 'xls'}. Default is 3.}

\item{xls_max_w}{an integer indicating the maximum window size for the
\code{'xls'} method. Only used if \code{algorithm = 'xls'}. Default is 15.}
}
\value{
a list containing:
\itemize{
\item{\code{ncomp}: the number of PLS components.}
\item{\code{coefficients}: the matrix of regression coefficients.}
\item{\code{bo}: a matrix containing the intercepts for each component.}
\item{\code{X_loadings}: the matrix of X loadings.}
\item{\code{Y_loadings}: the matrix of Y loadings.}
\item{\code{projection_mat}: the projection matrix for computing scores
from new data.}
\item{\code{transf}: a list containing:
\itemize{
\item{\code{Xcenter}: row vector of column means used for centering.}
\item{\code{Xscale}: row vector of column standard deviations used
for scaling (ones if \code{scale = FALSE}).}
}
}
\item{\code{weights}: the matrix of PLS weights.}
}
}
\description{
Computes orthogonal scores partial least squares (PLS)
regression using either NIPALS or SIMPLS algorithm. Supports multiple
response variables. In contrast to \code{opls}, this function omits
auxiliary outputs (e.g. scores, explained variance) not required for
local regression. For internal use only.
}
\references{
de Jong, S. (1993). SIMPLS: An alternative approach to partial least
squares regression. Chemometrics and Intelligent Laboratory Systems,
18(3), 251-263.

Shenk, J.S., & Westerhaus, M.O. (1991). Populations structuring of near
infrared spectra and modified partial least squares regression. Crop
Science, 31(6), 1548-1555.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
