% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_stratified.R
\name{optim_sample_strata}
\alias{optim_sample_strata}
\title{A function to construct an optimal strata for the samples, based on
the distribution of the given y.}
\usage{
optim_sample_strata(y, n)
}
\arguments{
\item{y}{a matrix of one column with the response variable.}

\item{n}{number of samples that must be sampled.}
}
\value{
a list with two \code{data.table} objects: \code{sample_strata} contains
the optimal strata, whereas \code{samples_to_get} contains information on how
many samples per stratum are supposed to be drawn.
}
\description{
for internal use only! This function computes the optimal strata
from the distribution of the given y
}
\keyword{internal}
