% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ospls}
\alias{ospls}
\title{orthogonal scores algorithn of partial leat squares (ospls)}
\usage{
ospls(X, Y, ncomp, scale)
}
\arguments{
\item{X}{a \code{matrix} of predictor variables.}

\item{Y}{a \code{matrix} of a single response variable.}

\item{ncomp}{the number of PLS components.}

\item{scale}{logical indicating whether \code{Xz} must be scaled.}

\item{center}{logical indicating whether \code{Xz} must be centered.}
}
\value{
a list containing the following elements:
\itemize{
\item{\code{coefficients}}{ the \code{matrix} of regression coefficients.}
\item{\code{bo}}{ a \code{matrix} of one row containing the intercepts for each component.}
\item{\code{scores}}{ the \code{matrix} of scores.}
\item{\code{X.loadings}}{ the \code{matrix} of X loadings.}
\item{\code{Y.loadings}}{ the \code{matrix} of Y loadings.}
\item{\code{projectionM}}{ the projection \code{matrix}.}
\item{\code{variance}}{ a \code{list} conating two objects: \code{x.var} and \code{y.var}. 
These objects contain information on the explained variance for the \code{X} and \code{Y} matrices respectively.}
\item{\code{transf}}{ a \code{list} conating two objects: \code{Xcenter} and \code{Xscale}. 
These objects contain information on the explained variance for the \code{X} and \code{Y} matrices respectively.}
}
}
\description{
Computes a partial least square (PLS) model for a single response variable. 
For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

