% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ospls2}
\alias{ospls2}
\title{orthogonal scores algorithn of partial leat squares 2 (ospls2)}
\usage{
ospls2(X, Y, ncomp, center, scale, maxiter, tol)
}
\arguments{
\item{X}{a \code{matrix}.}

\item{Y}{a \code{matrix}.}

\item{ncomp}{the number of PLS components.}

\item{center}{logical indicating whether \code{X} must be centered.}

\item{scale}{logical indicating whether \code{X} must be scaled.}

\item{maxiter}{maximum number of iterations.}

\item{tol}{limit for convergence of the algorithm in the nipals algorithm (default is 1e-06).}
}
\value{
a list containing the following elements:
\itemize{
\item{\code{scores}}{ the \code{matrix} of scores.}
\item{\code{X.loadings}}{ the \code{matrix} of X loadings.}
\item{\code{Y.loadings}}{ the \code{matrix} of Y loadings.}
\item{\code{projectionM}}{ the projection \code{matrix}.}
\item{\code{variance}}{ a \code{list} conating two objects: \code{x.var} and \code{y.var}. 
These objects contain information on the explained variance for the \code{X} and \code{Y} matrices respectively.}
}
}
\description{
Computes a partial least square (PLS) model for multiple reponse variables. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

