% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overall_var}
\alias{overall_var}
\title{Function for computing the overall variance of a matrix}
\usage{
overall_var(X)
}
\arguments{
\item{X}{a matrix.}
}
\value{
a vector of standard deviation values.
}
\description{
Computes the variance of a matrix. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
