% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pgpcv_cpp}
\alias{pgpcv_cpp}
\title{Internal Cpp function for performing leave-group-out cross validations for pls regression}
\usage{
pgpcv_cpp(X, Y, mindices, pindices, noisev = 0.001, scale)
}
\arguments{
\item{X}{a \code{matrix} of predictor variables.}

\item{Y}{a \code{matrix} of a single response variable.}

\item{mindices}{a \code{matrix} with \code{n} rows and \code{m} columns where \code{m} is equivalent to the number of
resampling iterations. The elements of each column indicate the indices of the samples to be used for modeling at each
iteration.}

\item{pindices}{a \code{matrix} with \code{k} rows and \code{m} columns where \code{m} is equivalent to the number of
resampling iterations. The elements of each column indicate the indices of the samples to be used for predicting at each
iteration.}

\item{scale}{a logical indicating whether both the predictors
and the response variable must be scaled to zero mean and unit variance.}

\item{mindices}{a \code{matrix} with \code{n} rows and \code{m} columns where \code{m} is equivalent to the number of
resampling iterations. The elements of each column indicate the indices of the samples to be used for modeling at each
iteration.}

\item{pindices}{a \code{matrix} with \code{k} rows and \code{m} columns where \code{m} is equivalent to the number of
resampling iterations. The elements of each column indicate the indices of the samples to be used for predicting at each
iteration.}

\item{ncomp}{an integer indicating the number of pls components.}
}
\value{
a list containing the following one-row matrices:
\itemize{
\item{\code{rmse.seg}}{ the RMSEs.}
\item{\code{st.rmse.seg}}{ the standardized RMSEs.}
\item{\code{rsq.seg}}{ the coefficients of determination.}
}
}
\description{
For internal use only!.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

