% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gesearch.R
\name{plot.gesearch}
\alias{plot.gesearch}
\title{Plot method for an object of class \code{gesearch}}
\usage{
\method{plot}{gesearch}(x, g = c("rmse", "n"), ...)
}
\arguments{
\item{x}{an object of class \code{gesearch} (as returned by \code{gesearch}).}

\item{g}{a character vector indicating what results shall be plotted.
Options are: \code{"rmse"} (for plotting the progress of the maximum RMSE
found during the iterations) or \code{"n"} (for plotting the cumulative
number of observations removed at each iteration).}

\item{...}{some arguments to be passed to the plot methods.}
}
\description{
Plots objects of class \code{gesearch}
}
\seealso{
\code{\link{gesearch}}
}
\author{
Leonardo Ramirez-Lopez
}
