% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rslocal.R
\name{plot.rslocal}
\alias{plot.rslocal}
\title{Plot method for an object of class \code{rslocal}}
\usage{
\method{plot}{rslocal}(x, g = c("rmse", "n"), ...)
}
\arguments{
\item{x}{an object of class \code{rslocal} (as returned by \code{rslocal}).}

\item{g}{a character vector indicating what results shall be plotted.
Options are: \code{"rmse"} (for plotting the progress of the maximum RMSE found
during the iterations) and/or \code{"n"} (for plotting the cumulative number
of observations removed at each iteration).}

\item{...}{some arguments to be passed to the plot methods.}
}
\description{
Plots objects of class \code{rslocal}
}
\seealso{
\code{\link{rslocal}}
}
\author{
Leonardo Ramirez-Lopez
}
