% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_helpers.R
\name{pls_cv}
\alias{pls_cv}
\title{Cross validation for PLS regression}
\usage{
pls_cv(
  x,
  y,
  ncomp,
  method = c("pls", "wapls"),
  center = TRUE,
  scale,
  min_component = 1,
  new_x = matrix(0, 1, 1),
  weights = NULL,
  p = 0.75,
  number = 10,
  group = NULL,
  retrieve = TRUE,
  tune = TRUE,
  max_iter = 1,
  tol = 1e-06,
  seed = NULL,
  modified = FALSE
)
}
\description{
for internal use only!
}
\keyword{internal}
