% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl.R
\name{plsCv}
\alias{plsCv}
\title{Cross validation for PLS regression}
\usage{
plsCv(x, y, ncomp, center = TRUE, scaled, weights = NULL, p = 0.75,
  resampling = 10, group = NULL, retrieve = c("final.model", "all.models",
  "none"))
}
\description{
for internal use only!
}
\keyword{internal}

