% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pplscv_cpp}
\alias{pplscv_cpp}
\title{Internal Cpp function for performing leave-group-out cross validations for pls regression}
\usage{
pplscv_cpp(X, Y, mindices, pindices, ncomp, scale)
}
\arguments{
\item{X}{a \code{matrix} of predictor variables.}

\item{Y}{a \code{matrix} of a single response variable.}

\item{mindices}{a \code{matrix} with \code{n} rows and \code{m} columns where \code{m} is equivalent to the number of 
resampling iterations. The elements of each column indicate the indices of the samples to be used for modeling at each 
iteration.}

\item{pindices}{a \code{matrix} with \code{k} rows and \code{m} columns where \code{m} is equivalent to the number of 
resampling iterations. The elements of each column indicate the indices of the samples to be used for predicting at each 
iteration.}

\item{ncomp}{an integer indicating the number of pls components.}

\item{scale}{a logical indicating whether the matrix of predictors must be scaled.}
}
\value{
a list containing the following one-row matrices:
\itemize{
\item{\code{rmse.seg}}{ the RMSEs.}
\item{\code{st.rmse.seg}}{ the standardized RMSEs.}
\item{\code{rsq.seg}}{ the coefficients of determination.}
}
}
\description{
For internal use only!.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

