% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mbl.R
\name{pred.gpr.dp}
\alias{pred.gpr.dp}
\title{Prediction function for "gpr.dp" (Gaussian process regression with dot product covariance)}
\usage{
pred.gpr.dp(object, newdata)
}
\description{
internal
}
\keyword{internal}

