% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_opls}
\alias{predict_opls}
\title{Prediction function for the \code{opls} and \code{fopls} functions}
\usage{
predict_opls(bo, b, ncomp, newdata, scale, Xscale)
}
\arguments{
\item{bo}{a numeric value indicating the intercept.}

\item{b}{the matrix of regression coefficients.}

\item{ncomp}{an integer value indicating how may components must be used in the prediction.}

\item{newdata}{a matrix containing the predictor variables.}

\item{scale}{a logical indicating whether the matrix of predictors used to create the regression model was scaled.}

\item{Xscale}{if \code{scale = TRUE} a matrix of one row with the values that must be used for scaling \code{newdata}.}
}
\value{
a matrix of predicted values.
}
\description{
Predicts response values based on a model generated by either by \code{opls} or the \code{fopls} functions.
For internal use only!.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
