% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predopls}
\alias{predopls}
\title{Prediction function for the \code{opls} and \code{opls2} functions}
\usage{
predopls(bo, b, ncomp, newdata, scale, Xscale)
}
\arguments{
\item{bo}{a numeric value indicating the intercept.}

\item{b}{the \code{matrix} of regression coefficients.}

\item{ncomp}{an integer value indicating how may components must be used in the prediction.}

\item{newdata}{a \code{matrix} containing the predictor variables.}

\item{scale}{a logical indicating whether the matrix of predictors used to create the regression model 
(either in \code{opls} or \code{opls2}) was scaled.}

\item{Xscale}{if \code{scale = TRUE} a \code{matrix} of one row with the values that must be used for scaling \code{newdata}.}
}
\value{
a \code{matrix} of predicted values
}
\description{
Predicts response values based on a model generated by either by \code{opls} or the \code{opls2} functions. 
For internal use only!.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

