% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gesearch.R
\name{print.gesearch}
\alias{print.gesearch}
\title{Print method for an object of class \code{gesearch}}
\usage{
\method{print}{gesearch}(x, ...)
}
\arguments{
\item{x}{an object of class \code{gesearch} (as returned by the \code{gesearch} function).}

\item{...}{arguments to be passed to methods (not functional).}
}
\description{
Prints the content of an object of class \code{gesearch}
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
