% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.local_fit.R
\name{print.local_fit}
\alias{print.local_fit}
\title{Print method for an object of class \code{local_fit}}
\usage{
\method{print}{local_fit}(x, ...)
}
\arguments{
\item{x}{an object of class \code{local_fit}}

\item{...}{not yet functional.}
}
\description{
Prints the contents of an object of class \code{local_fit}
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
