% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.local_ortho_diss.R
\name{print.local_ortho_diss}
\alias{print.local_ortho_diss}
\title{Print method for an object of class \code{ortho_diss}}
\usage{
\method{print}{local_ortho_diss}(x, ...)
}
\arguments{
\item{x}{an object of class \code{localortho_diss} (returned by
\code{ortho_diss} when it uses \code{.local = TRUE}).}

\item{...}{arguments to be passed to methods (not yet functional).}
}
\description{
Prints the content of an object of class \code{ortho_diss}
}
\author{
Leonardo Ramirez-Lopez and Antoine Stevens
}
\keyword{internal}
