\name{print.localOrthoDiss}
\alias{print.localOrthoDiss}
\title{Print method for an object of class \code{orthoDiss}}
\usage{
  \method{print}{localOrthoDiss}(x, ...)
}
\arguments{
  \item{x}{an object of class \code{localOrthoDiss} (which
  is returned by \code{orthoDiss} when it uses \code{local
  = TRUE}).}

  \item{...}{arguments to be passed to methods (not yet
  functional).}
}
\description{
  Print method for an object of class \code{orthoDiss}
}
\author{
  Leonardo Ramirez-Lopez and Antoine Stevens
}

