% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mbl.R
\name{print.mbl}
\alias{print.mbl}
\title{Print method for an object of class \code{mbl}}
\usage{
\method{print}{mbl}(x, ...)
}
\arguments{
\item{x}{an object of class \code{mbl} (as returned by the \code{mbl} function).}

\item{...}{arguments to be passed to methods (not functional).}
}
\description{
Prints the content of an object of class \code{mbl}
}
\author{
Leonardo Ramirez-Lopez and Antoine Stevens
}
\keyword{internal}
