% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ortho_projection.R
\name{print.ortho_projection}
\alias{print.ortho_projection}
\title{Print method for an object of class \code{ortho_projection}}
\usage{
\method{print}{ortho_projection}(x, ...)
}
\arguments{
\item{x}{an object of class \code{ortho_projection} (as returned by the 
\code{ortho_projection} function).}

\item{...}{arguments to be passed to methods (not yet functional).}
}
\description{
Prints the contents of an object of class \code{ortho_projection}
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
