% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.rslocal.R
\name{print.rslocal}
\alias{print.rslocal}
\title{Print method for an object of class \code{rslocal}}
\usage{
\method{print}{rslocal}(x, ...)
}
\arguments{
\item{x}{an object of class \code{rslocal} (as returned by the \code{rslocal} function).}

\item{...}{arguments to be passed to methods (not functional).}
}
\description{
Prints the content of an object of class \code{rslocal}
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
