% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reconstruction_error}
\alias{reconstruction_error}
\title{Projection to pls and then re-construction}
\usage{
project_opls(projection_mat, ncomp, newdata, scale, Xcenter, Xscale)
}
\arguments{
\item{x}{a matrix to project.}

\item{projection_mat}{the projection matrix generated by the \code{opls_get_basics} function.}

\item{xloadings}{the loadings matrix generated by the \code{opls_get_basics} function.}
}
\value{
a matrix of 1 row and 1 column.
}
\description{
Projects spectra onto a PLS space and then reconstructs it back.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
