% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslocal.R
\name{rslocal}
\alias{rslocal}
\alias{rslocal.default}
\alias{rslocal.formula}
\alias{predict.rslocal}
\title{Data-driven search and optimization of sample-specific pls models (rslocal)}
\usage{
\method{rslocal}{formula}(formula, train, test, 
        k, b, .keep = 0.95, 
        method = "pls",
        optimize = "reconstruction",
        pls_c, pls_tune = FALSE, 
        rs_ctrl = rs_control(),
        group = NULL,
        documentation = character())

\method{rslocal}{default}(Xr, Yr, Xu, Yu = NULL, 
        k, b, .keep = 0.95, 
        method = "pls",
        optimize = "reconstruction",
        pls_c, pls_tune = FALSE, 
        rs_ctrl = rs_control(),
        group = NULL,
        documentation = character())

\method{predict}{rslocal}(object, newdata, type = 'response', ...)
}
\arguments{
\item{formula}{(BETTER DESCRIPTION REQUIRED) a \code{formula}}

\item{train}{(BETTER DESCRIPTION REQUIRED) a \code{data.frame}}

\item{test}{a (BETTER DESCRIPTION REQUIRED) a \code{data.frame}}

\item{Xr}{a \code{matrix} of predictor variables of the spectral library (observations in rows and variables in columns)}

\item{Yr}{a numeric \code{vector} containing the values of the response variable corresponding to the Xr data}

\item{Xu}{a \code{matrix} of predictor variables of the 'local' or site-specific samples (observations in rows and variables in columns)}

\item{Yu}{a numeric \code{vector} containing the values of the response variable corresponding to the Xu data. Default is \code{NULL}}

\item{k}{(BETTER DESCRIPTION REQUIRED) the number of TRAIN samples randomly selected in the re-sampling step, and also the target number of SL samples returned by the algorithm}

\item{b}{(BETTER DESCRIPTION REQUIRED) the number of times each sample in the TRAIN should be tested, on average, in each iteration of the algorithm}

\item{.keep}{a numeric value larger than 0 and below 1 (default 0.95). This value indicates:
 \itemize{
 \item{if the \code{...$keep_by} parameter of the object passed to the \code{rs_ctrl} argument is \code{'proportion'}:}{ a percentage of the samples to be kept at each iteration. The proportion of samples to be removed is \code{1 - .keep}. See \code{keep_by} argument of the \code{\link{rs_control}} function.} 
 \item{if the \code{...$keep_by} parameter of the object passed to the \code{rs_ctrl} argument is \code{'probability'}:}{ a probability value to be used to estimate the percentile (cut point of the distribution) of associated errors. In this case samples with associated errors below this estimated percentile value are kept and the ones equal to or above this value are removed. See \code{keep_by} argument of the \code{\link{rs_control}} function.} 
}}

\item{method}{a character string indicating the method to be used for the internal calibrations within the rs-local alrgorithm. The only method available for the moment is \code{"pls"}.}

\item{optimize}{(BETTER DESCRIPTION REQUIRED). CONDUCT THE SEARCH BASED ON THE SPECTRAL RECONSTRUCTION ERROR OR THE RMSE?. Options are: \code{'response'} or \code{'reconstruction'} (default).}

\item{pls_c}{the number of pls components to be used in the selected regression method.}

\item{pls_tune}{an optional boolean \code{value} to specify the method for pls factor selection when \code{"pls"} is used. If TRUE a 10-fold cross validation is used to inform selection and pls.c is ignored. if \code{FALSE} the number of factors must be specified using pls.c. Setting to TRUE greatly increases processing time. Default is FALSE.}

\item{rs_ctrl}{a list created with the \code{\link{rs_control}} function which contains additional parameters that further control some aspects of the \code{rs_local} function. The default list is as returned by \code{rs_control()}. See the \code{\link{rs_control}} function for more details.}

\item{group}{an optional \code{factor} (or \code{vector} that can be coerced to a \code{factor} by \code{as.factor}) that assigns to each sample in the training set (i.e. \code{train} or \code{Xr}) a group/class label (e.g. groups can be given by spectra collected from the same batch of measurements, from the same sample, from samples with very similar origin, etc). This is taken into account for internal leave-group-out cross validation for pls tuning (factor optimization) to avoid pseudo-replication. When one observation is selected for cross validation, all observations of the same group are removed together and assigned to validation. The length of the \code{vector} must be equal to the number of samples in the training set (i.e. \code{nrow(train)} or \code{nrow(Xr)}). See details.}

\item{documentation}{(BETTER DESCRIPTION REQUIRED) an optional character string for documentating the call to this function.}
}
\value{
a \code{list} (BETTER ANDUPDATED DESCRIPTION REQUIRED) with the following components:
\itemize{
 \item{\code{Kx}:}{ a \code{data.frame} of predictor variables of the specific samples selected for 'local' or site specific calibration}
 \item{\code{Ky}:}{ a numeric \code{vector} containing the values of the response variable corresponding to the K.x data}
 \item{\code{idx}:}{ a numeric \code{vector} containing the index into the original SL (row numbers) for the selected samples (K)}
 \item{\code{iter_rmse}:}{ BETTER DESCRIPTION REQUIRED a \code{data.frame} with the maximum associated error found for the kept at the end of each iteration (\code{max_rmse} and the associated error above which the ramples where removed at each iteration (column \code{cut_rmse}). If the samples kept at each iteration were retained based on a given probability (specified in the \code{.keep} argument), the \code{cut_rmse} column indicates the cut-off percentile value.}
 \item{\code{n_removed}:}{ a \code{data.frame} with the number of samples removed at the end of each iteration.}
 \item{\code{validation_results}:}{ DESCRIBE}
 \item{\code{rs_ctrl}:}{ a list mirroring the one provided in the \code{rs_ctrl} argument.}
 \item{\code{final_model}:}{ DESCRIBE}
 \item{\code{documentation}:}{ a character string mirroring the one provided in the \code{documentation} argument.}
 }
}
\description{
This function implements the resampling-local (RS-LOCAL) algorithm which conducts a search of a suitable calibration subset from a large dataset based on a given set of samples
}
\details{
The \code{RS-LOCAL} algorithm requires a large datase (where the sample search is conducted), a subset of specific samples, \code{m}, 
and three parameters, \code{k}, \code{b} and \code{r}, which are described below. The \code{m} samples should be 
measured with the spectrometer and analyzed using the reference laboratory analytical method. 
These samples should also be representative of the entire population from the study site. They may be selected, 
for example, from all of the study site's spectra using the Kennard-Stone (KS) 
method (Kennard & Stone, 1969). \code{RS-LOCAL} uses the \code{m} data and re-sampling to evaluate and then keep the relevant samples and remove irrelevant ones
from the SL so that the data that remain are the most appropriate for deriving a site-specific calibration.
The SL subset, \code{K}, that remains is typically augmented by the user with the \code{m} data by spiking. Together, \code{K} plus \code{m} 
 form an \code{RS-LOCAL} data set, which can be used for deriving site-specific spectroscopic calibrations.
When selecting values for the \code{k},\code{b},\code{.keep} parameters,
\itemize{
 \item{\code{k:}}{ The size of the sampled dataset used in the internal calibration and validation, and the target number of SL samples ultimately returned by the algorithm. For recommended values see Lobsey et al. 2017.}
 \item{\code{r:}}{ The number of times each training sample is tested in each iteration of the algorithm (on average). More consistent results are achieved with high \code{b} values, however this will increase processing time. A recommended value for \code{b} is greater than 40.}
 \item{\code{.keep:}}{ This determines the number of samples to be kept at each iteration of the algorithm. More consistent results are achieved with large \code{.keep} values, however this will increase processing time. It is recomended to use \code{.keep} values larger than 0.9.}
 }
}
\examples{
\dontrun{
#' ## GOOD EXAMPLES ARE REQUIRED
}
}
\references{
Lobsey C., Viscarra Rossel R.A., Pierre R., Hedley C.B. 2017. RS-LOCAL data-mines information from large spectral libraries to improve local calibrations. European Journal of Soil Science.

Kennard, R.W. & Stone, L.A. 1969. Computer aided design of experiments. Technometrics, 11(1), pp.137-148.
}
\author{
Craig Lobsey, Raphael Viscarra Rossel and Leonardo Ramirez-Lopez
}
