% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.local_ortho_diss.R
\name{[.local_ortho_diss}
\alias{[.local_ortho_diss}
\title{Print method for an object of class \code{local_ortho_diss}}
\usage{
\method{[}{local_ortho_diss}(x, rows, columns, drop = FALSE, ...)
}
\arguments{
\item{x}{\code{local_ortho_diss} matrix}

\item{rows}{the indices of the rows}

\item{columns}{the indices of the columns}

\item{drop}{drop argument}

\item{...}{not used}
}
\description{
prints the subsets of localortho_diss objects
}
\keyword{internal}
