% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{waplsoneweights_cpp}
\alias{waplsoneweights_cpp}
\title{Internal Cpp function for computing the weights of the PLS components necessary for weighted average PLS}
\usage{
waplsoneweights_cpp(projectionm, 
xloadings, 
coefficients, 
newX, 
minF, 
maxF, 
scale, 
Xcenter, 
Xscale)
}
\arguments{
\item{projectionm}{the projection \code{matrix} generated either by the \code{ospls} or \code{ospls2}.}

\item{xloadings}{.}

\item{coefficients}{the \code{matrix} of regression coefficients.}

\item{newX}{a \code{matrix} of one new spectra to be predicted.}

\item{minF}{an integer indicating the minimum number of pls components.}

\item{maxF}{an integer indicating the maximum number of pls components.}

\item{scale}{a logical indicating whether the matrix of predictors used to create the regression model 
(either in \code{ospls} or \code{ospls2}) was scaled.}

\item{Xcenter}{a \code{matrix} of one row with the values that must be used for centering \code{newdata}.}

\item{Xscale}{if \code{scale = TRUE} a \code{matrix} of one row with the values that must be used for scaling \code{newdata}.}
}
\value{
a \code{matrix} of one row with the weights for each component between the max. and min. specified.
}
\description{
For internal use only!.
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}

