% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wcolSds}
\alias{wcolSds}
\title{Function for identifiying the column in a \code{matrix} with the largest standard deviation}
\usage{
wcolSds(X)
}
\arguments{
\item{X}{a \code{matrix}.}
}
\value{
a value indicating the index of the column with the largest standard deviation.
}
\description{
Identifies the column with the largest standard deviation. For internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
