% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_min_vector}
\alias{which_min_vector}
\title{A function to compute indices of minimum values of a distance vector}
\usage{
which_min_vector(X)
}
\arguments{
\item{X}{a vector of distances}
}
\value{
a vector of the indices of the nearest neighbors
}
\description{
For internal use only
}
\details{
Used internally to find the nearest neighbors.
It searches in lower (or upper) triangular matrix. Therefore this must be the format of the
input data. The piece of code int \code{len = (sqrt(X.size()*8+1)+1)/2} generated an error in CRAN
since \code{sqrt} cannot be applied to integers.
}
\author{
Antoine Stevens
}
\keyword{internal}
