% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslocal_helpers.R
\name{biter}
\alias{biter}
\title{fits pls models at each r-local iteration}
\usage{
biter(
  itersubs,
  Xu,
  Yu,
  iter_sequence,
  optimization,
  ncomp,
  tune,
  p,
  number,
  scale,
  max_iter,
  tol,
  allow_parallel
)
}
\description{
internal
}
\keyword{internal}
