% Generated by roxygen2 (4.0.1): do not edit by hand
\name{e2m}
\alias{e2m}
\title{A function for transforming a matrix from its Euclidean spcae to its Mahalanobis space}
\usage{
e2m(X, sm.method = c("svd", "eigen"))
}
\description{
A function for transforming a matrix from its Euclidean spcae to its Mahalanobis space
}
\keyword{internal}

