% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fastDistVVL}
\alias{fastDistVVL}
\title{A fast (serial) algorithm of (squared) Euclidean cross-distance for vectors written in C++}
\usage{
fastDistVVL(X)
}
\arguments{
\item{X}{a \code{vector}}
}
\value{
a \code{vector} of distance (lower triangle of the distance matrix, stored by column)
}
\description{
A fast (parallel for linux) algorithm of (squared) Euclidean cross-distance for vectors written in C++
}
\details{
used internally in orthoProjection
}
\author{
Leo Ramirez-Lopez
}
\keyword{internal}
