% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{final_fits_cpp}
\alias{final_fits_cpp}
\title{Compute Final Local PLS Model Outputs}
\usage{
final_fits_cpp(X, Y, new_x, min_component, max_component, scale, maxiter, tol)
}
\arguments{
\item{X}{A matrix of predictor variables used for calibration.}

\item{Y}{A matrix of response variables used for calibration.}

\item{new_x}{A single observation (1 x p) of predictor variables to compute weights.}

\item{min_component}{The minimum number of PLS components to include in the final model.}

\item{max_component}{The maximum number of PLS components to include in the final model.}

\item{scale}{Logical indicating whether to scale the data.}

\item{maxiter}{Maximum number of iterations allowed during the NIPALS algorithm.}

\item{tol}{Tolerance threshold for convergence in the iterative algorithm.}
}
\value{
A list with the following elements:
\itemize{
\item \code{ib0}: Intercept term, computed as a weighted sum of component-specific intercepts.
\item \code{ibs}: Weighted regression coefficients.
\item \code{ivips}: Weighted variable importance (VIP) scores, scaled by component SDs.
\item \code{isratio}: Weighted selectivity ratios, scaled by component SDs.
\item \code{Xscale}: Scaling vector used to scale \code{X} (if \code{scale = TRUE}).
}
}
\description{
This function calculates the final local model coefficients, intercept,
scaled variable importance (VIP), and selectivity ratio for a single observation
using a weighted combination of PLS components.
}
\keyword{internal}
