% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_library.R
\name{fit_library}
\alias{fit_library}
\title{Build predictive/fucntional libraries}
\usage{
fit_library(Xr, Yr, sm, k, k_diss, k_range, ws = NULL, pls_c, 
            pc_selection = list("opc", min(dim(Xr), 40)),
            group, build_around, gh = TRUE, center = TRUE,
            scale = FALSE, diss_predictors = FALSE,
            metric = "rmse", return_best = TRUE,
            pls_max_iter = 1, pls_tol = 1e-6,
            documentation = character(), ...)
}
\arguments{
\item{Xr}{a numeric matrix (or data.frame) of predictor variables of dimentions \code{n * p} corresponding to the reference data (observations in rows and variables in columns).}

\item{Yr}{a numeric vector of length \code{n} containing the values of the response variable corresponding to the reference data. Missing values might be allowed (see details).}

\item{diss_method}{Either a character string indicating the dissimilarity
metric to be used for selecting the nearest neighbors, or a precomputed
dissimilarity matrix.

If a character is provided, it must be one of the following:
\itemize{
\item \code{"pca"}: Mahalanobis distance based on principal component (PC) scores
computed via singular value decomposition (SVD). See \code{\link[=ortho_diss]{ortho_diss()}}.
\item \code{"pca.nipals"}: Same as \code{"pca"} but using the NIPALS algorithm. See
\code{\link[=ortho_diss]{ortho_diss()}}.
\item \code{"pls"}: Mahalanobis distance based on PLS scores. Requires \code{Yr}. See
\code{\link[=ortho_diss]{ortho_diss()}}.
\item \code{"mpls"}: Mahalanobis distance based on modified PLS scores (Shenk and
Westerhaus, 1991; Westerhaus, 2014). Requires \code{Yr}. See \code{\link[=ortho_diss]{ortho_diss()}}.
\item \code{"cor"}: Dissimilarity based on correlation between observations.
See \code{\link[=cor_diss]{cor_diss()}}.
\item \code{"euclid"}: Euclidean distance. See \code{\link[=f_diss]{f_diss()}}.
\item \code{"cosine"}: Cosine distance. See \code{\link[=f_diss]{f_diss()}}.
\item \code{"sid"}: Spectral information divergence. See \code{\link[=sid]{sid()}}.
}

These options are passed to the \code{\link[=dissimilarity]{dissimilarity()}} function, which performs
the actual computation. Additional arguments to \code{dissimilarity()} can be
passed through \code{...}.

If a matrix is provided, it must represent a precomputed dissimilarity
structure with constraints depending on whether \code{anchor_indices} is used:
\itemize{
\item If \code{anchor_indices = NULL} (default), \code{diss_method} must be a square
matrix of size \verb{n x n}, where \code{n = nrow(Xr)}. It must be symmetric
with zeros on the diagonal, representing self-dissimilarity.
\item If \code{anchor_indices} is specified (with length \code{m}), then
\code{diss_method} must have dimensions \verb{n x m}, where \code{n = nrow(Xr)} and
\code{m = length(anchor_indices)}. The submatrix \code{diss_method[anchor_indices, ]}
must be symmetric with zeros on the diagonal.
}

In all cases, the dissimilarity matrix must be consistent with the order
of samples in \code{Xr}. That is, the rows of \code{diss_method} must correspond
exactly to the rows of \code{Xr}, and when used, the elements of
\code{anchor_indices} must refer to valid rows in both \code{Xr} and \code{diss_method}.}

\item{k}{a vector of integers specifying the sequence of k nearest neighbours to be tested. Either \code{k} or \code{k_diss} must be specified. This \code{vector} will be automatically sorted into ascending order. The values of vectors of class numeric, will be coerced to the next upper inetegers.}

\item{k_diss}{NOT YET IMPLEMENTED a numeric \code{vector} specifying the sequence of dissimilarity thresholds to be tested for the selection of the nearest neighbors around each observation. For a given observation, its neighbors are those that exhibit a dissimilarity value equal to or below a given threshold. These thresholds depend on the corresponding dissimilarity measure specified in \code{sm}. Either \code{k} or \code{k_diss} must be specified.}

\item{k_range}{NOT YET IMPLEMENTED a vector of two integers specifying the minimum (first value) and the maximum (second value) number of neighbours allowed when the \code{k_diss} argument is used.}

\item{ws}{an optional odd integer value which specifies a window size when the correlation dissimilarity is used (i.e \code{sm = "cor"}). If not specified, the standard correlation dissimilarity is computed (i.e no moving window). Default is \code{NULL}.}

\item{pls_c}{a vector of two integers indicating the minimum (first value) and maximum (second value) number of pls components to be used in the weighted average pls (wapls) regressions.}

\item{pc_selection}{a list specifying the details of the method to be used for identifying the number of components to be retained for computing the dissimilarities between samples when the method passed to \code{sm} is one of the following options: \code{"pc"}, \code{"loc.pc"}, \code{ "pls"} or \code{ "loc.pls"}. This list must contain two elements in the following order: \itemize{
\item{\code{method}:}{the method for selecting the number of components. Possible options are:  \code{"opc"} (optimized pc selection based on Ramirez-Lopez et al. (2013a, 2013b). See the \code{\link{orthoProjection}} function for more details;  \code{"cumvar"} (for selecting the number of principal components based on a given cumulative amount of explained variance); \code{"var"} (for selecting the number of principal components based on a given amount of explained variance); and  \code{"manual"} (for specifying manually the desired number of principal components)}
\item{\code{value}:}{a numerical value that complements the selected method. If \code{"opc"} is chosen, it must be a value indicating the maximal number of principal components to be tested (see Ramirez-Lopez et al., 2013a, 2013b). If \code{"cumvar"} is chosen, it must be a value (larger than 0 and below 1) indicating the maximum amount of cumulative variance that the retained components should explain. If \code{"var"} is chosen, it must be a value (larger than 0 and below 1) indicating that components that explain (individually) a variance lower than this threshold must be excluded. If \code{"manual"} is chosen, it must be a value specifying the desired number of principal components to retain.
}}
The default method for the \code{pc_selection} argument is \code{"opc"} and the maximum number of principal components to be tested is set to \code{min(dim(Xr), 40)}, where code{Xr} is the matrix of reference predictors in the \code{\link{mbl}} function.
Optionally, the \code{pc_selection} argument admits \code{"opc"} or \code{"cumvar"} or \code{"var"} or \code{"manual"} as a single character string. In such a case the default for \code{"value"} when either \code{"opc"} or \code{"manual"} are used is \code{min(dim(Xr), 40)}. When \code{"cumvar"} is used the default \code{"value"} is set to 0.99 and when \code{"var"} is used the default \code{"value"} is set to 0.01.}

\item{group}{(PUT THE DETAILS OF THE 1_NN VALIDATION IN DETAILS) an optional \code{factor} (or \code{vector} that can be coerced to a \code{factor} by \code{as.factor}) that assigns to each observation in \code{Xr} a group/class label (e.g. groups can be given by spectra collected from the same batch of measurements, from the same sample, from samples with very similar origin, etc). This is taken into account for internal validation of pls models (and factor optimization) to avoid pseudo-replication. For validation, the model of a given target observation is first fitted based on its neighbors and excluding that observation (and all the samples belonging to its group), then the model is used to predict the response variable of the target observation. See details.}

\item{anchor_indices}{An optional integer vector of length \code{m} (\code{m < n}) specifying the row indices of the reference samples around which local models should be constructed. If \code{NULL} (default), local models will be built for all \code{n} samples in the reference set. See Details.}

\item{gh}{a logical indicating whether or not to compute and return the Mahalanobis distance (in the pls space) between each element in \code{Xr} and the center of \code{Xr}.}

\item{center}{a logical indicating whether or not the predictor variables must be centered at each local segment (before regression).}

\item{scale}{a logical indicating whether or not the predictor variables must be scaled to unit variance at each local segment (before regression).}

\item{diss_predictors}{a logical indicating if the local (square symmetric) dissimilarity matrix corresponding the selected neighbourhood shall be used as source of additional predictors (i.e the columns of this local matrix are treated as predictor variables). In some cases this may result in an improvement  of the prediction performance (Ramirez-Lopez et al., 2013a).}

\item{metric}{a character value indicating what model perfoamance statistics shall be used to select the best set of parameters (i.e. number of neighbors or threshold distances and pls factors)  to fit the final model. Options are \code{"rmse"} (default) or \code{"r2"} (coefficient of determination).}

\item{return_best}{a logical indicating if the final library of functions using the optimal parameters found shall be returned.}

\item{pls_max_iter}{(BETTER DESCRIPTION REQUIRED) maximum number of iterations for the partial least squares methods.}

\item{pls_tol}{(BETTER DESCRIPTION REQUIRED) for convergence in the partial orthogonal scores partial least squares regressions using the nipals algorithm. Default is 1e-6}

\item{chunk_size}{Integer. Specifies the maximum number of samples
(rows or columns, depending on context) to include in each chunk.
Each chunk is processed independently, which can be beneficial
for parallel computation. For example, each chunk can be dispatched
to a separate processor or core, where the samples within the chunk
are processed sequentially. Defaults to \code{1} (i.e., one sample per
chunk). The value of \code{chunk_size} cannot exceed the total number of
available samples for the given context (e.g., \code{nrow(Xr)} or
\code{length(anchor_indices)} when provided).}

\item{documentation}{(BETTER DESCRIPTION REQUIRED) an optional character string for documentating the call to this function.}

\item{...}{arguments passed to the \code{\link{dissimilarity}} function.}
}
\value{
DESCRIOTION REQUIRED
}
\description{
This function builds a library of predictive functions based on memory-based learning
}
\details{
** Anchor indices **

By default, local models are constructed around all \code{n} samples in the
reference set. Alternatively, the user may specify a subset of \code{m} samples
(\code{m < n}) around which the library of local models should be built. This is
done by supplying their row indices to the \code{anchor_indices} argument.

In this configuration, each local model is still constructed using the
nearest neighbors selected from the full set of \code{n} reference samples, but
only for the \code{m} anchor samples. This approach is especially useful when the
reference set is very large, as it reduces the total number of models while
retaining representativeness and predictive performance—particularly if the
anchor samples are chosen to reflect the diversity of the dataset.

This strategy can lead to substantial computational savings and a more
compact library of models. \strong{Note} that when distance calculations rely on
the response variable (e.g., GH distance, PLS distance, or optimized PC
distance), the response values (\code{Yr}) of the anchor samples themselves are
not used during distance computation. This is done to improve computational
efficiency: dissimilarities are computed between each anchor sample and all
reference samples (excluding the anchor itself), rather than recalculating
the full \verb{n × n} distance matrix. This significantly reduces the number of
response-based distance computations when the number of anchors is small.
Also, \strong{note} that the response values (Yr) of the anchor observations are
always used during the fitting of  their corresponding local models.
For efficiency, the number of anchor indices must be less than 90\% of the
total number of rows in \code{Xr}. If this threshold is exceeded, consider
building a model for each sample in \code{Xr} by setting \code{anchor_indices = NULL}.

** Missing values in \code{Yr} **

Missing values in \code{Yr} are allowed. If they exceed 25\% of the total number
of observations, a warning will be issued. The local model of an observation
with a missing \code{Yr} value is fitted only using its nearest neighbors (i.e.,
without including the observation itself).
}
\examples{
\dontrun{
#' ## GOOD EXAMPLES ARE REQUIRED
}
}
\references{
Rajalahti, T., Arneberg, R., Berven, F. S., Myhr, K. M., Ulvik, R. J., &
Kvalheim, O. M. (2009). Biomarker discovery in mass spectral profiles by
means of selectivity ratio plot. Chemometrics and Intelligent Laboratory
Systems, 95(1), 35-48.
}
\author{
Leonardo Ramirez-Lopez
}
