% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl.R
\name{gpr.dp}
\alias{gpr.dp}
\title{Gaussian process regression with dot product covariance}
\usage{
gpr.dp(Xn, Yn, noise.v = 0.001, scaled = TRUE, center = FALSE, x.add = NULL)
}
\description{
internal
}
\keyword{internal}
