% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mbl.R
\name{gprCv}
\alias{gprCv}
\title{Cross validation for Gaussian process regression}
\usage{
gprCv(x, y, scaled, weights = NULL, p = 0.75, resampling = 10,
  group = NULL, noise.v = 0.001, retrieve = c("final.model", "none"))
}
\description{
internal
}
\keyword{internal}

