% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortho_diss_helpers.R
\name{ith_subsets_ortho_diss}
\alias{ith_subsets_ortho_diss}
\title{iterator for nearest neighbor subsets}
\usage{
ith_subsets_ortho_diss(x, xu = NULL, y, kindx, na_rm = FALSE)
}
\arguments{
\item{x}{a reference matrix}

\item{xu}{a second matrix}

\item{y}{a matrix of side information}

\item{kindx}{a matrix of nearest neighbor indices}

\item{remove}{NA indices from kindx (useful when working with dissimilarity
thresholds to select the neighbors (instead of fixed number of neighbors).}
}
\description{
internal
}
\keyword{internal}
