\name{locFit}
\alias{locFit}
\title{Local multivariate regression}
\usage{
locFit(x, y, predMethod, scaled = TRUE, weights = NULL, pred.new = TRUE,
  newdata, pls.c, CV = FALSE, resampling = 10, p = 0.75,
  w2_warning = TRUE, noise.v = 0.001, range.pred.lim = TRUE)
}
\description{
Local multivariate regression
}
\keyword{internal}

