% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl.R
\name{locFit}
\alias{locFit}
\title{Local multivariate regression}
\usage{
locFit(x, y, predMethod, scaled = TRUE, weights = NULL, pred.new = TRUE,
  newdata, pls.c, CV = FALSE, resampling = 10, p = 0.75, group = NULL,
  noise.v = 0.001, range.pred.lim = TRUE, pls.max.iter = 1,
  pls.tol = 1e-06)
}
\description{
internal
}
\keyword{internal}

