% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{opls}
\alias{opls}
\title{orthogonal scores algorithn of partial leat squares (opls)}
\usage{
opls(X, 
     Y, 
     ncomp, 
     scale, 
     maxiter, 
     tol)
}
\arguments{
\item{X}{a matrix of predictor variables.}

\item{Y}{a matrix of either a single or multiple response variables.}

\item{ncomp}{the number of pls components.}

\item{scale}{logical indicating whether \code{X} must be scaled.}

\item{maxiter}{maximum number of iterations.}

\item{tol}{limit for convergence of the algorithm in the nipals algorithm.}

\item{(for}{weights computation): xls_max_w an integer indicating the maximum window size for the "xls"
method. Only used if \code{algorithm = 'xls'}. Default is 15 (as in BUCHI NIRWise PLUS software).}
}
\value{
a list containing the following elements:
\itemize{
\item{\code{coefficients}}{ the matrix of regression coefficients.}
\item{\code{bo}}{ a matrix of one row containing the intercepts for each component.}
\item{\code{scores}}{ the matrix of scores.}
\item{\code{X_loadings}}{ the matrix of X loadings.}
\item{\code{Y_loadings}}{ the matrix of Y loadings.}
\item{\code{projection_mat}}{ the projection matrix.}
\item{\code{Y}}{ the \code{Y} input.}
\item{\code{transf}}{ a \code{list} conating two objects: \code{Xcenter} and \code{Xscale}}. 
\item{\code{weights}}{ the matrix of wheights.}}
}
\description{
Computes orthogonal socres partial least squares (opls) 
regressions with the NIPALS algorithm. It allows multiple response variables. 
It does not return the variance information of the components. NOTE: For 
internal use only!
}
\author{
Leonardo Ramirez-Lopez
}
\keyword{internal}
