% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ortho_projection.R
\name{plot.ortho_projection}
\alias{plot.ortho_projection}
\title{Plot method for an object of class \code{ortho_projection}}
\usage{
\method{plot}{ortho_projection}(x, col = "dodgerblue", ...)
}
\arguments{
\item{x}{an object of class \code{ortho_projection} (as returned by \code{ortho_projection}).}

\item{col}{the color of the plots (default is "dodgerblue")}

\item{...}{arguments to be passed to methods.}
}
\description{
\lifecycle{maturing}

Plots objects of class \code{ortho_projection}
}
\seealso{
\code{\link{ortho_projection}}
}
\author{
Leonardo Ramirez-Lopez and Antoine Stevens
}
