% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mbl.R
\name{plsCv}
\alias{plsCv}
\title{Standard deviation of columns
#}
\usage{
plsCv(x, y, ncomp, method = c("pls", "wapls1"), center = TRUE, scaled,
  minF = 1, newX = matrix(0, 1, 1), weights = NULL, p = 0.75,
  resampling = 10, group = NULL, retrieve = c("final.model", "all.models",
  "none"), max.iter = 1, tol = 1e-06)
}
\description{
internal
#
}
\details{
#
}
\keyword{internal}

