% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plsCv}
\alias{plsCv}
\title{Cross validation for PLS regression}
\usage{
plsCv(x, y, ncomp, scaled, weights, p = 0.75, resampling = 10,
  group = NULL, retrieve = c("final.model", "all.models", "none"))
}
\description{
Cross validation for PLS regression
}
\keyword{internal}

