% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbl.R
\name{plsCv}
\alias{plsCv}
\title{Cross validation for PLS regression}
\usage{
plsCv(x, y, ncomp, method = c("pls", "wapls1"), center = TRUE, scaled,
  minF = 1, newX = matrix(0, 1, 1), weights = NULL, p = 0.75,
  resampling = 10, group = NULL, retrieve = TRUE, .optimize = TRUE,
  max.iter = 1, tol = 1e-06)
}
\description{
for internal use only!
}
\keyword{internal}

