\name{pred.gpr.dp}
\alias{pred.gpr.dp}
\title{Prediction function for "gpr.dp" (Gaussian process regression with dot product covariance)}
\usage{
pred.gpr.dp(object, newdata)
}
\description{
Prediction function for "gpr.dp" (Gaussian process
regression with dot product covariance)
}
\keyword{internal}

