% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.localOrthoDiss.R
\name{print.localOrthoDiss}
\alias{print.localOrthoDiss}
\title{Print method for an object of class \code{orthoDiss}}
\usage{
\method{print}{localOrthoDiss}(x, ...)
}
\arguments{
\item{x}{an object of class \code{localOrthoDiss} (returned by \code{orthoDiss} when it uses \code{local = TRUE}).}

\item{...}{arguments to be passed to methods (not yet functional).}
}
\description{
Prints the content of an object of class \code{orthoDiss}
}
\author{
Leonardo Ramirez-Lopez and Antoine Stevens
}

