\name{print.mbl}
\alias{print.mbl}
\title{Print method for an object of class \code{mbl}}
\usage{
  \method{print}{mbl}(x, ...)
}
\arguments{
  \item{x}{an object of class \code{mbl} (as returned by
  \code{mbl}).}

  \item{...}{arguments to be passed to methods (not yet
  functional).}
}
\description{
  Print method for an object of class \code{mbl}
}
\author{
  Leonardo Ramirez-Lopez and Antoine Stevens
}

