% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resemble.R
\docType{package}
\name{resemble-package}
\alias{resemble-package}
\alias{resemble}
\title{Overview of the functions in the resemble package}
\description{
\lifecycle{maturing}

Functions for memory-based learning
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
}
\details{
This is the version \code{2.0} (\code{'gordillo'}) of the package. It
implements a number of \code{R} functions useful for
modeling complex spectral spectra (e.g. NIR, IR).
The package includes functions for dimensionality reduction,
computing spectral dissimilarity matrices, nearest neighbor search,
and modeling spectral data using memory-based learning.

The functions available for dimensionality reduction are:
\itemize{
  \item{\code{\link{ortho_projection}}}
  \item{\code{\link{pc_projection}}}
  \item{\code{\link{pls_projection}}}
  \item{\code{\link{predict.ortho_projection}}}
  }
The functions available for computing dissimilarity matrices are:
\itemize{
  \item{\code{\link{dissimilarity}}}
  \item{\code{\link{f_diss}}}
  \item{\code{\link{cor_diss}}}
  \item{\code{\link{sid}}}
  \item{\code{\link{ortho_diss}}}
  }
The functions available for evaluating dissimilarity matrices are:
\itemize{
  \item{\code{\link{sim_eval}}}
  }
The functions available for nearest neighbor search:
\itemize{
  \item{\code{\link{search_neighbors}}}
  }
The functions available for modeling spectral data:
\itemize{
  \item{\code{\link{mbl_control}}}
  \item{\code{\link{mbl}}}
  \item{\code{\link{rs_control}}}
  \item{\code{\link{rslocal}}}
  \item{\code{\link{predict.rslocal}}}
  }
Other supplementary functions:
\itemize{
  \item{\code{\link{plot.mbl}}}
  \item{\code{\link{plot.rslocal}}}
  \item{\code{\link{plot.ortho_projection}}}
  }
}
\author{
LLL
}
