% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rslocal.R
\name{rslocal}
\alias{rslocal}
\alias{rslocal.default}
\alias{rslocal.formula}
\alias{predict.rslocal}
\title{Data-driven search and optimization in spectral libraries for 
building site-specific calibrations (RS-LOCAL)}
\usage{
\method{rslocal}{formula}(formula, train, test,
        k, b, method, 
        ..., na_action = na.pass)
        
\method{rslocal}{default}(Xr, Yr, Xu, Yu = NULL,
        k, b, retain = 0.95,
        method = local_fit_pls(pls_c = min(dim(Xr), 10)),
        optimization = "reconstruction",
        control = rs_control(),
        group = NULL,
        scale = FALSE,
        documentation = character(), ...)
        
\method{predict}{rslocal}(object, newdata, type = 'response', ...)
}
\arguments{
\item{...}{optional parameters used in method `formula` to be passed to the 
low level function rslocal.default. Not currently used for `predict` and 
`default` methods of `rslocal`.}

\item{formula}{an object of class \link[stats]{formula} which represents the 
basic model to be use.}

\item{train}{a data.frame with the training data containing the variables in
the model.}

\item{test}{a data.frame with the test data (local observations) containing 
the variables in the model.}

\item{Xr}{a matrix of predictor variables of the reference data 
(observations in rows and variables in columns).}

\item{Yr}{a matrix of one column containing the values of the 
response variable corresponding to the reference data.}

\item{Xu}{a matrix of predictor variables of the 'local' or 
site-specific observations (observations in rows and variables in columns).}

\item{Yu}{a matrix of one column containing the values of the response 
variable corresponding to the Xu data. It is only mandatory if 
\code{optimization = "response"}. Default is \code{NULL}.}

\item{k}{the number of reference/training observations selected in 
the re-sampling step of the rs-local algorithm. 
See description.}

\item{b}{the number of times each observation in the training set (\code{Xr}) 
should be tested, on average, in each iteration of the rs-local 
algorithm. See description.}

\item{retain}{a numeric value larger than 0 and below 1 (default 0.95). 
This value indicates:
 \itemize{
 \item{if the \code{...$retain_by} parameter of the object passed to the 
 \code{control} argument is \code{'proportion'}:}{ a percentage of the 
 observations to be kept at each iteration. The proportion of observations to be 
 removed is \code{1 - retain}. See \code{retain_by} argument of the 
 \code{\link{rs_control}} function.}
 \item{if the \code{...$retain_by} parameter of the object passed to the 
 \code{control} argument is \code{'probability'}:}{ a probability value to be 
 used to estimate the percentile (cut point of the distribution) of 
 associated errors. In this case, observations with associated errors below this 
 estimated percentile value are kept and the ones equal to or above this 
 value are removed. See \code{retain_by} argument of the 
 \code{\link{rs_control}} function.}
}}

\item{method}{an object of class \code{\link{local_fit}} which indicates the 
type of regression to conduct within the rs-local alrgorithm as well as 
additional parameters affecting this regression. See \code{\link{local_fit}} 
function. The only method allowed for the moment is pls, i.e. methods created 
with \code{local_fit_pls()}.}

\item{optimization}{a character string indicating the sample search method. 
Options are: 
\itemize{
 \item{\code{'response'}: }{The observations are retained based on the root 
 mean squared error of the prediction of the response variable in the test 
 set (\code{Yu}). In this case, it is required to pass the response values 
 for the test set (\code{Yu}).}
 \item{\code{'reconstruction'} (Default): }{The observations are retained 
 based on the spectral reconstruction error estimated for the test set 
 (\code{Xu}). This error is estimated by projecting the test set onto the pls 
 space (using the pls model built in each iteration with a fixed number of 
 pls factors) and then back-transforming (reconstruct) the projected test set 
 to its spectral space. Finally, the RMSE of this reconstruction is computed 
 and used to select the observations. This reconstruction is done with a fixed 
 number of pls components for all the iterations so that reconstruction errors 
 are comparable. In this case, it response values for the test set are not 
 required.}
 }}

\item{control}{a list created with the \code{\link{rs_control}} function 
which contains additional parameters that further control some aspects of the 
\code{rs_local} function. The default list is as returned by 
\code{rs_control()}. See the \code{\link{rs_control}} function for more 
details.}

\item{scale}{a logical indicating if the predictor variables must be scaled 
to unit variance at each iteration before regression.}

\item{group}{an optional factor (or vector that can be coerced 
to \code{\link[base]{factor}} by \code{as.factor}) that assigns to each observation in 
the training set (i.e. \code{train} or \code{Xr}) a group/class label (e.g. 
groups can be given by spectra collected from the same batch of measurements, 
from the same observation, from observations with very similar origin, etc). 
This is taken into account for internal leave-group-out cross validation for 
pls tuning (factor optimization) to avoid pseudo-replication. When one 
observation is selected for cross validation, all observations of the same 
group are removed together and assigned to validation. The length of the 
vector must be equal to the number of observations in the training set 
(i.e. \code{nrow(train)} or \code{nrow(Xr)}). See details.}

\item{documentation}{an optional character string that can be used to 
describe anything related to the \code{mbl} call (e.g. description of the 
input data). Default: \code{character()}. NOTE: his is an experimental 
argument.}

\item{na_action}{a function to specify the action to be taken if NAs are 
found in the data passed to train. Default \code{\link[stats]{na.fail}}. 
NOTE: \code{na_action} is not applied to data passed to \code{test}. If given, 
this argument must be named.}

\item{object}{an object of class `rslocal`, as that created by the function 
\code{\link{rslocal}}.}

\item{newdata}{a data frame or matrix containing new data.}

\item{type}{a character vector indicating what to return. Options are: 
`response` (default) and `scores`.}
}
\value{
a \code{list} with the following elements:
\itemize{
 \item{\code{x_local}:}{ a matrix of predictor variables corresponding to the 
 observations selected.}
 \item{\code{y_local}:}{ a matrix of one column with the response variable 
 corresponding to the observations selected.}
 \item{\code{indices}:}{ a numeric vector with the indices of the 
 observations selected from the original training set.}
 \item{\code{iter_rmse}:}{ a data.table with the maximum associated error 
 found for the observations retained at the end of each iteration 
 (\code{max_rmse}) and the associated error above which the observations 
 where removed at each iteration (column \code{cut_rmse}). If the 
 observations were retained based on a given probability (specified 
 in the \code{retain} argument), the \code{cut_rmse} column indicates the 
 cut-off percentile value.}
 \item{\code{n_removed}:}{ a data.table with the number of observations 
 removed at the end of each iteration.}
 \item{\code{validation_results}:}{ a list containing some validation results. 
 This list has two elements:
 \itemize{
 \item{\code{val_info}:}{ a list containing the indices of the observations 
 (in the original training set) in each re-sampling iteration used for the 
 validation of the final model. A matrix with the predictions for the test set 
 using the selected   observations is also returned.}
 \item{\code{results}:}{ a list containing the results of the 
 leave-group-out validation done for the selected training observations and 
 if \code{Yr} was supplied, the validation results for the test set.}
 }
 }
 \item{\code{control}:}{ a list mirroring the one provided in the 
 \code{control} argument.}
 \item{\code{final_model}:}{ a list with the following elements of a pls 
 regression model:
 #'  \itemize{
 \item{\code{npls}:}{ The number of pls factors used.}
 \item{\code{coefficients}:}{ The regression coefficients for each factor.}
 \item{\code{bo}:}{ The intercepts of each factor.}
 \item{\code{scores}:}{ The matrix of pls scores.}
 \item{\code{X_loadings}:}{ The matrix of pls loadings for the predictor 
 variables.}
 \item{\code{Y_loadings}:}{ The matrix of pls loadings for the response 
 variable.}
 \item{\code{projection_mat}:}{ The matrix for pls projections.}
 \item{\code{vip}:}{ The matrix of variable importance for projection of 
 each factor.}
 \item{\code{selectivity_ratio}:}{ The matrix of variable importance for each 
 factor based on the method of selectivity ratio (Rajalahti et al., 2009).}
 \item{\code{Y}:}{ A matrix with the response values used to fit the final 
 pls model.}
 \item{\code{weights}:}{ A matrix of pls weights.}
 }
 }
 \item{\code{documentation}:}{ a character string mirroring the one provided 
 in the \code{documentation} argument.}
 }
}
\description{
\lifecycle{maturing}

This function implements the re-sampling local (RS-LOCAL) algorithm. The 
algorithm selects a subset from a large calibration data set that is more 
appropriate for deriving 'local' or site-specific calibrations.
It uses a data-driven approach to capture the local or site-specific 
relationship between the response and predictor variables often not 
represented or is generalized by existing large calibration data sets 
(e.g. spectral libraries).
}
\details{
The rs-local algorithm requires a large data set (where the sample 
search is conducted), a subset of specific observations, \code{m},
and three parameters, \code{k}, \code{b} and \code{r}, which are described 
below. The predictor variables of the \code{m} observations must be available, 
along with (optionally) their response values.

These observations should also be representative of the entire population 
which they originated from. They may be selected, for example, from a large 
set of observations (with unknown response values) using sampling method such 
Kennard-Stone (Kennard & Stone, 1969). 

The rs-local algorithm uses the \code{m} data and  re-sampling to evaluate 
and then retain the relevant observations and remove irrelevant ones from 
the reference training set so that the data that remain 
are the most appropriate for deriving a data-specific regression model.
If response values are available for observations in the test set, these are 
used to augment the subset of the reference set returned by the rs-local 
algorithm to finally build the data-specific regression model.
When selecting values for the \code{k},\code{b},\code{retain} parameters,
\itemize{
 \item{\code{k:}}{ The size of the randomly sampled data set used in the 
 internal calibration and validation step of the algorithm. It is also the 
 target number of reference/training observations returned by the algorithm 
 i.e. when insufficient  samples remain in the reference/training set to 
 continue re-sampling. For recommended values see 
 Lobsey et al. 2017.}
 \item{\code{r:}}{ The number of times each training observation is tested in each 
 iteration of the algorithm (on average). More consistent results are 
 achieved with high \code{b} values, however this will increase processing 
 time. A recommended value for \code{b} is greater than 40.}
 \item{\code{retain:}}{ This determines the number of observations to be kept at 
 each iteration of the algorithm. More consistent results are achieved with 
 large \code{retain} values, however this will increase processing time. It 
 is recommended to use \code{retain} values larger than 0.9.}
 }
}
\examples{
\dontrun{
#' ## GOOD EXAMPLES ARE REQUIRED

## could we base an example on LUCAS (SL) and NIRsoil (local) - these are 
Belgium, are they in LUCAS?

## First example showing basic implementation, returning index (k_idx) 
in the the SL, then spiking (to final rslocal calibration data set)

## Second example showing use of final model / prediction (using internal 
functions)

## Third example showing use of method 'reconstruction'
}
}
\references{
Lobsey, C. R., Viscarra Rossel, R. A., Roudier, P., & Hedley, C. B. 2017. 
rs-local data-mines information from spectral libraries to improve local 
calibrations. European Journal of Soil Science, 68(6), 840-852.

Kennard, R.W. & Stone, L.A. 1969. Computer aided design of experiments. 
Technometrics, 11(1), pp.137-148.

Rajalahti, T., Arneberg, R., Berven, F. S., Myhr, K. M., Ulvik, R. J., 
Kvalheim, O. M. 2009. Biomarker discovery in mass spectral profiles by means 
of selectivity ratio plot. Chemometrics and Intelligent Laboratory Systems, 
95(1), 35-48.
}
\author{
Craig Lobsey, Raphael Viscarra Rossel and Leonardo Ramirez-Lopez
}
