% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mbl.R
\name{wapls.weights}
\alias{wapls.weights}
\title{Internal function for computing the weights of the PLS components necessary for weighted average PLS}
\usage{
wapls.weights(plsO, orgX, type = "w1", newX = NULL, pls.c)
}
\arguments{
\item{plsO}{either an object returned by the \code{plsCv} function or an object as returned by the \code{opls} function which contains a pls model.}

\item{orgX}{the original spectral \code{matrix} which was used for calibrating the pls model.}

\item{type}{type of weight to be computed. The only available option (for the moment) is \code{"w1"}. See details on the \code{mbl} function where it is explained how \code{"w1"} is computed whitin the \code{"wapls1"} regression.}

\item{newX}{a \code{vector} of a new spectral sample. When "w1" is selected, newX must be specified.}

\item{pls.c}{a \code{vector} of length 2 which contains both the minimum and maximum number of PLS components for which the weights must be computed.}
}
\value{
\code{wapls.weights} returns a \code{vector} of weights for each PLS component specified
}
\description{
internal
}
\author{
Leonardo Ramirez-Lopez and Antoine Stevens
}
\keyword{internal}

