% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_min}
\alias{which_min}
\title{A function to compute row-wise index of minimum values of a square distance matrix}
\usage{
which_min(X)
}
\arguments{
\item{X}{a square matrix of distances}
}
\value{
a vector of the indices of the minimum value in each row of the input matrix
}
\description{
For internal use only
}
\details{
Used internally to find the nearest neighbors
}
\author{
Antoine Stevens
}
\keyword{internal}
