% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{which_minV}
\alias{which_minV}
\title{A function to compute indices of minimum values of a distance vector}
\usage{
which_minV(X,cores)
}
\arguments{
\item{X}{a \code{vector} of distance (as computed in \code{resemble:::fastDistVV} or \code{base::dist})}

\item{cores}{number of cores used to run the computation}
}
\value{
a \code{vector} of the indices of the nearest neighbours
}
\description{
For internal use only
}
\details{
Used internally to find the nearest neighbours.
It searches in lower (or upper?) trianguular matrix. Therefore this must be the format of the
input data. The piece of code int \code{len = (sqrt(X.size()*8+1)+1)/2} generated an error in CRAN
since \code{sqrt} cannot be applied to integers.
}
\author{
Antoine Stevens
}
\keyword{internal}

